Separated Account tab logic
